Agent Windy Game
Once upon a time there was a good king in a peaceful kingdom. One day a
witch came to this
kingdom and imprisoned the princess (daughter of the king) in a cave.
Then the witch ensured the
security of the cave by placing a dragon in front of the cave door.
At that time, there was a super hero called Agent Windy. The King made a
request from Agent
Wendy to survive the princess from the witch. Agent Wendy accepted the
challenge and went to
save the princess.
Agent Wendy travels through a road that full of obstacles. It is
required to go to the cave within 3
minutes. The dragon sends fire balls towards the Agent Wendy. Agent
Wendy needs to shoot the
fireballs to destroy them. Initially Agent has 5 lives. If fireballs
touch the Agent, the lives will be
reduced by one. Fairies drop healthy mushrooms from the sky. If the
Agent Wendy catches 10 of
them, his life will be refilled by one. Same time the witch sends
poisonous mushrooms. If the Agent
Wendy catches a poisonous mushroom his journey will be delayed by 10
seconds. Finally Agent
needs to reach to the dragon and shoot 15 bullets to kill the dragon.
Then the Agent Wendy can win
the challenge and save the princess.